// cosmwasm/tools/client/TokenBridge.types.ts
export type Binary = string;
export interface InstantiateMsg {
  chain_id: number;
  gov_address: Binary;
  gov_chain: number;
  native_decimals: number;
  native_denom: string;
  native_symbol: string;
  wormhole_contract: string;
  wrapped_asset_code_id: number;
  [k: string]: unknown;
}